backend/requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy[asyncio]
aiosqlite
pydantic
pydantic_settings
asyncpg
psycopg[binary]
openai
python-jose[cryptography]
passlib[bcrypt]